requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-WTF==1.0.1
gitdb==4.0.10
GitPython==3.1.30
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.24.1
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
